Create EvaluateRecommender.java
In this program, we first load data from .CSV file using interface called Datamodel.  It is a user-based recommender.  we find the similarity between users using PearsonCorrelation similarity. Now, we define which similar users are to be taken as neighbors. create a recommender using the data model, similarity and neighborhood.